requirements.txt
streamlit
torch
torchvision
pillow
fastapi
uvicorn
python-multipart
